# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Notebook API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  color: string;
  createdAt: string;
  updatedAt: string;
  pageCount: number;
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  color?: string;
}

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  color?: string;
}

export interface Page {
  id: number;
  bookId: number;
  title: string;
  content: string;
  pageNumber: number;
  createdAt: string;
  updatedAt: string;
}

export interface PageInput {
  /** @minLength 1 */
  title: string;
  content?: string;
}

export interface PageUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
}

export interface Summary {
  totalBooks: number;
  totalPages: number;
  recentBooks: Book[];
}