# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Notebook API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: books
    description: Book operations
  - name: pages
    description: Page operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

  /books/{bookId}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a single book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book and all its pages
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /books/{bookId}/pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all pages in a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new page in a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "201":
          description: Page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"

  /books/{bookId}/pages/{pageId}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a single page
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Page not found
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update a page
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: Page updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Page not found
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /summary:
    get:
      operationId: getSummary
      tags: [books]
      summary: Get notebook summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Book:
      type: object
      required: [id, title, color, createdAt, updatedAt, pageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        pageCount:
          type: integer

    BookInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        color:
          type: string

    BookUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        color:
          type: string

    Page:
      type: object
      required: [id, bookId, title, content, pageNumber, createdAt, updatedAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        title:
          type: string
        content:
          type: string
        pageNumber:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    PageInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string

    PageUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string

    Summary:
      type: object
      required: [totalBooks, totalPages, recentBooks]
      properties:
        totalBooks:
          type: integer
        totalPages:
          type: integer
        recentBooks:
          type: array
          items:
            $ref: "#/components/schemas/Book"