# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Notebook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  /** @minLength 1 */
  title: string;
  color?: string;
}