# lib/api-zod/src/generated/types/page.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Notebook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Page {
  id: number;
  bookId: number;
  title: string;
  content: string;
  pageNumber: number;
  createdAt: string;
  updatedAt: string;
}