# lib/api-zod/src/generated/types/pageUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Notebook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PageUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
}